Remove not needed apis
