--- MoviesExample/app/src/main/java/com/novoda/demo/movies/MoviesListBinding.java
@@ -1,16 +1,0 @@
-package com.novoda.demo.movies;
-
-import android.databinding.BindingAdapter;
-import android.support.v7.widget.RecyclerView;
-
-public class MoviesListBinding {
-
-    @BindingAdapter({"app:items"})
-    public static void setItems(RecyclerView recyclerView, MoviesSate moviesSate) {
-        MoviesAdapter adapter = (MoviesAdapter) recyclerView.getAdapter();
-        if (adapter != null && moviesSate != null) {
-            adapter.setMoviesSate(moviesSate);
-        }
-    }
-
-}
